+++ packages/core-app-api/src/apis/implementations/OAuthRequestApi/OAuthRequestManager.ts
@@ -16,9 +16,9 @@
 
 import {
   OAuthRequestApi,
-  PendingAuthRequest,
-  AuthRequester,
-  AuthRequesterOptions,
+  PendingOAuthRequest,
+  OAuthRequester,
+  OAuthRequesterOptions,
 } from '@backstage/core-plugin-api';
 import { Observable } from '@backstage/types';
 import { OAuthPendingRequests, PendingRequest } from './OAuthPendingRequests';
@@ -34,11 +34,17 @@ import { BehaviorSubject } from '../../../lib/subjects';
  * @public
  */
 export class OAuthRequestManager implements OAuthRequestApi {
-  private readonly subject = new BehaviorSubject<PendingAuthRequest[]>([]);
-  private currentRequests: PendingAuthRequest[] = [];
+  private readonly subject = new BehaviorSubject<PendingOAuthRequest[]>([]);
+  private currentRequests: PendingOAuthRequest[] = [];
   private handlerCount = 0;
 
-  createAuthRequester<T>(options: AuthRequesterOptions<T>): AuthRequester<T> {
+  createAuthRequester<T>(options: OAuthRequesterOptions<T>): OAuthRequester<T> {
+    if (!options.provider.id) {
+      // eslint-disable-next-line no-console
+      console.warn(
+        'DEPRECATION WARNING: Not passing a provider id to createAuthRequester is deprecated, it will be required in the future',
+      );
+    }
     const handler = new OAuthPendingRequests<T>();
 
     const index = this.handlerCount;
@@ -67,8 +73,8 @@ export class OAuthRequestManager implements OAuthRequestApi {
   // Converts the pending request and popup options into a popup request that we can forward to subscribers.
   private makeAuthRequest(
     request: PendingRequest<any>,
-    options: AuthRequesterOptions<any>,
-  ): PendingAuthRequest | undefined {
+    options: OAuthRequesterOptions<any>,
+  ): PendingOAuthRequest | undefined {
     const { scopes } = request;
     if (!scopes) {
       return undefined;
@@ -88,7 +94,7 @@ export class OAuthRequestManager implements OAuthRequestApi {
     };
   }
 
-  authRequest$(): Observable<PendingAuthRequest[]> {
+  authRequest$(): Observable<PendingOAuthRequest[]> {
     return this.subject;
   }
 }