+++ plugins/azure-devops/src/index.ts
@@ -23,3 +23,18 @@ export {
 } from './plugin';
 
 export { AzurePullRequestsIcon } from './components/AzurePullRequestsIcon';
+
+export { FilterType } from './components/PullRequestsPage';
+export type {
+  PullRequestColumnConfig,
+  BaseFilter,
+  Filter,
+  PullRequestFilter,
+  AssignedToUserFilter,
+  CreatedByUserFilter,
+  AssignedToTeamFilter,
+  CreatedByTeamFilter,
+  AssignedToTeamsFilter,
+  CreatedByTeamsFilter,
+  AllFilter,
+} from './components/PullRequestsPage';