+++ plugins/search-backend-module-elasticsearch/src/engines/index.ts
@@ -15,4 +15,7 @@
  */
 
 export { ElasticSearchSearchEngine } from './ElasticSearchSearchEngine';
-export type { ConcreteElasticSearchQuery } from './ElasticSearchSearchEngine';
+export type {
+  ConcreteElasticSearchQuery,
+  ElasticSearchClientOptions,
+} from './ElasticSearchSearchEngine';